Move asset source related code from Media.Browser to Media

+++ Neos.Media/Classes/Command/MediaCommandController.php
@@ -18,11 +18,14 @@ use Neos\Flow\Annotations as Flow;
 use Neos\Flow\Cli\CommandController;
 use Neos\Flow\Persistence\PersistenceManagerInterface;
 use Neos\Flow\ResourceManagement\PersistentResource;
+use Neos\Media\Domain\Model\Asset;
 use Neos\Media\Domain\Model\AssetInterface;
+use Neos\Media\Domain\Model\AssetSource\AssetSourceAwareInterface;
 use Neos\Media\Domain\Repository\AssetRepository;
 use Neos\Media\Domain\Repository\ThumbnailRepository;
 use Neos\Media\Domain\Service\ThumbnailService;
 use Neos\Media\Domain\Strategy\AssetModelMappingStrategyInterface;
+use Neos\Utility\Files;
 
 /**
  * @Flow\Scope("singleton")
@@ -137,51 +140,86 @@ class MediaCommandController extends CommandController
     /**
      * Remove unused assets
      *
-     * This command iterates over all existing assets, checks their usage count
-     * and lists the assets which are not reported as used by any AssetUsageStrategies.
-     * The unused assets can than be removed.
+     * This command iterates over all existing assets, checks their usage count and lists the assets which are not
+     * reported as used by any AssetUsageStrategies. The unused assets can than be removed.
      *
+     * @param string $assetSource If specified, only assets of this asset source are considered. For example "neos" or "my-asset-management-system"
+     * @param bool $quiet If set, only errors will be displayed.
+     * @param bool $assumeYes If set, "yes" is assumed for the "shall I remove ..." dialogs
      * @return void
      */
-    public function removeUnusedCommand()
+    public function removeUnusedCommand(string $assetSource = '', bool $quiet = false, bool $assumeYes = false)
     {
         $iterator = $this->assetRepository->findAllIterator();
         $assetCount = $this->assetRepository->countAll();
         $unusedAssets = [];
-        $unusedAssetInfo = [];
+        $tableRowsByAssetSource = [];
         $unusedAssetCount = 0;
+        $unusedAssetsTotalSize = 0;
 
-        $this->outputLine('<b>Searching for unused assets:</b>');
-
-        $this->output->progressStart($assetCount);
-        /** @var AssetInterface $asset */
-        foreach ($this->assetRepository->iterate($iterator) as $asset) {
-            if ($asset->getUsageCount() === 0) {
-                $unusedAssets[] = $asset;
-                $unusedAssetInfo[] = sprintf('- %s (%s)', $asset->getIdentifier(), $asset->getResource()->getFilename());
-                $unusedAssetCount++;
-            }
-            $this->output->progressAdvance(1);
+        $filterByAssetSourceIdentifier = $assetSource;
+        if ($filterByAssetSourceIdentifier === '') {
+            !$quiet && $this->outputLine('<b>Searching for unused assets:</b>');
+        } else {
+            !$quiet && $this->outputLine('<b>Searching for unused assets of asset source "%s":</b>', [$filterByAssetSourceIdentifier]);
         }
+
+        !$quiet && $this->output->progressStart($assetCount);
+
+        foreach ($this->assetRepository->iterate($iterator) as $asset) {
+            if ($asset instanceof Asset && $asset->getUsageCount() === 0) {
+                if (!$asset instanceof AssetSourceAwareInterface) {
+                    continue;
+                }
+                if ($filterByAssetSourceIdentifier !== '' && $asset->getAssetSourceIdentifier() !== $filterByAssetSourceIdentifier) {
+                    continue;
+                }
+
+                $assetSource = $asset->getAssetSource();
+                $fileSize = str_pad(Files::bytesToSizeString($asset->getResource()->getFileSize()), 9, ' ', STR_PAD_LEFT);
+
+                $unusedAssets[] = $asset;
+                $tableRowsByAssetSource[$assetSource->getIdentifier()][] = [
+                    $asset->getIdentifier(),
+                    $asset->getResource()->getFilename(),
+                    $fileSize
+                ];
+                $unusedAssetCount++;
+                $unusedAssetsTotalSize += $asset->getResource()->getFileSize();
+            }
+            !$quiet && $this->output->progressAdvance(1);
+        }
+
+        !$quiet && $this->output->progressFinish();
 
         if ($unusedAssetCount === 0) {
-            $this->output->outputLine(PHP_EOL . sprintf('No unused assets found.', $unusedAssetCount));
-            $this->quit(0);
+            !$quiet && $this->output->outputLine(PHP_EOL . sprintf('No unused assets found.', $unusedAssetCount));
+            exit;
         }
 
-        $this->outputLine(PHP_EOL . 'Found the following unused assets: ' . PHP_EOL . implode(PHP_EOL, $unusedAssetInfo));
+        foreach ($tableRowsByAssetSource as $assetSourceIdentifier => $tableRows) {
+            !$quiet && $this->outputLine(PHP_EOL . 'Found the following unused assets from asset source <success>%s</success>: ' . PHP_EOL, [$assetSourceIdentifier]);
 
-        $continue = $this->output->askConfirmation(sprintf('Do you want to remove <b>%s</b> unused assets?', $unusedAssetCount));
-        if ($continue !== true) {
-            $this->quit(0);
+            !$quiet && $this->output->outputTable(
+                $tableRows,
+                ['Asset identifier', 'Filename', 'Size']
+            );
         }
 
-        $this->output->progressStart($unusedAssetCount);
+        !$quiet && $this->outputLine(PHP_EOL . 'Total size of unused assets: %s' . PHP_EOL, [Files::bytesToSizeString($unusedAssetsTotalSize)]);
+
+        if ($assumeYes === false) {
+            if (!$this->output->askConfirmation(sprintf('Do you want to remove <b>%s</b> unused assets?', $unusedAssetCount))) {
+                exit(1);
+            }
+        }
+
+        !$quiet && $this->output->progressStart($unusedAssetCount);
         foreach ($unusedAssets as $asset) {
-            $this->output->progressAdvance(1);
+            !$quiet && $this->output->progressAdvance(1);
             $this->assetRepository->remove($asset);
         }
-        $this->outputLine('');
+        !$quiet && $this->output->progressFinish();
     }
 
     /**
